VoltasBeko/Voltas_api_image/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.9.6
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.10.2
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
easyocr==1.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.3
fire==0.6.0
Flask==2.2.3
flask-babel==4.0.0
fonttools==4.50.0
fsspec==2024.3.1
future==1.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==2.0.3
pdf2docx==0.5.8
pillow==10.3.0
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.8
pybase64==1.3.2
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.8.1
rarfile==4.2
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
sympy==1.12
termcolor==2.4.0
tifffile==2023.7.10
torch==2.2.2+cu121
torchaudio==2.2.2+cu121
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
unicorn==2.0.1.post1
urllib3==2.2.1
uvicorn==0.29.0
visualdl==2.5.3
Werkzeug==2.2.3
zipp==3.18.1
